+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Hibernate annotations: classes annotated with a [`org.hibernate.annotations.GenericGenerator`](http://docs.jboss.org/hibernate/orm/4.2/manual/en-US/html/ch05.html#mapping-declaration-id-generator) that are referred by a class annotated with `javax.persistence.GeneratedValue` are recognized as live code.
 * Mark classes being annotated with the JAXB annotation `javax.xml.bind.annotation.XmlRegistry` as _live code_.
 * Added analysis of [Apache Tiles](http://tiles.apache.org) XML definition files: recognizing listed classes as _live code_.
+* Added analysis of a JEE6 feature: recognizing implementations of [`javax.servlet.ServletContainerInitializer`](http://docs.oracle.com/javaee/6/api/javax/servlet/ServletContainerInitializer.html) as _live code_.
 
 ## Internal
 * integrated logging in _tooling_ classes via [SLF4J](http://www.slf4j.org/)